# cmd/api/docs/swagger.yaml
definitions:
  domain.PaymentMethod:
    properties:
      maximumamount:
        type: number
      method_id:
        type: integer
      paymentmethod:
        type: string
    type: object
  req.LoginStruct:
    properties:
      password:
        maxLength: 64
        minLength: 8
        type: string
      username:
        maxLength: 12
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      description: Enter user_name  with password
      operationId: UserLogin
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/req.LoginStruct'
      responses:
        "200":
          description: successfully logged in
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: faild to generat JWT
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: api for user to login
      tags:
      - User Login
  /paymentmethod/add:
    post:
      description: Enter payment method and maximum amount
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentMethod'
      responses:
        "200":
          description: successfully  added payment method
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: failed to add payment method
          schema:
            $ref: '#/definitions/response.Response'
      summary: api for add payment method by admin
      tags:
      - Add payment method
  /paymentmethod/update:
    patch:
      description: Enter payment method and maximum amount with id
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentMethod'
      responses:
        "200":
          description: successfully  updated payment method
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: failed to updatepayment method
          schema:
            $ref: '#/definitions/response.Response'
      summary: api for update payment method by admin
      tags:
      - upadate payment method
  /paymentmethod/view:
    get:
      description: Get all products
      parameters:
      - description: Page
        format: int32
        in: query
        name: page
        required: true
        type: integer
      - description: Page Size
        format: int32
        in: query
        name: pagesize
        required: true
        type: integer
      responses:
        "200":
          description: successfully  displayed all prioducts
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: ferror while getting data
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all paymentmethods
      tags:
      - View all payment method
swagger: "2.0"